Implement proper socket agent refreshing
When the server asks the agent to refresh itself, it will then halt the
execution of the current script and wait for a new socket agent with
the same browser and page to connect. It then swaps the new agent’s
socket in for the old agent’s. As a result, refreshing can occur
without any other component having to be aware that the agent was
disconnected in the first place.